DM-4483:separated the draw webgrid and compute the grids in two files, changed codes based on Trey's suggestions
