+++ src/firefly/js/drawingLayers/WebGridUI.jsx
@@ -4,18 +4,12 @@
  * 4/15/16
  */
 import React, {PropTypes} from 'react';
-import {dispatchForceDrawLayerUpdate,dispatchWebGridLayerCoordinateChange} from '../visualize/DrawLayerCntlr.js';
+import {dispatchForceDrawLayerUpdate, dispatchModifyCustomField} from '../visualize/DrawLayerCntlr.js';
 import AppDataCntlr from '../core/AppDataCntlr.js';
-import {primePlot,getPlotViewById} from '../visualize/PlotViewUtil.js';
-import {SingleColumnMenu} from '../ui/DropDownMenu.jsx';
-import CoordinateSys from '../visualize/CoordSys.js';
-import lodash, {get} from 'lodash';
+import {get} from 'lodash';
 import {ListBoxInputFieldView} from '../ui/ListBoxInputField.jsx';
-import {COORDIANTE_PREFERENCE} from './WebGrid.js';
-import {
-    ToolbarButton,
-    DropDownVerticalSeparator} from '../ui/ToolbarButton.jsx';
-import {DropDownToolbarButton} from '../ui/DropDownToolbarButton.jsx';
+import {COORDIANTE_PREFERENCE} from './ComputeWebGridData.js';
+
 
 export const getUIComponent = (drawLayer,pv) => < WebGridUI drawLayer={drawLayer} pv={pv}/>;
 
@@ -37,16 +31,16 @@ function WebGridUI({drawLayer,pv}) {
     var pref= AppDataCntlr.getPreference(COORDIANTE_PREFERENCE);
    return  (
         <div>
-        <ListBoxInputFieldView
-        onChange={(request) => onCoordinateChange( pv.plotId,drawLayer, request) }
-        options={ coordinateOptionArray}
-        multiple={false}
-        value= {pref}
-        labelWidth={2}
-        label={''}
-        tooltip={'select a coordinate'}
-     />
-            </div>
+           <ListBoxInputFieldView
+               onChange={(request) => onCoordinateChange( pv.plotId,drawLayer, request) }
+               options={ coordinateOptionArray}
+               multiple={false}
+               value= {pref}
+               labelWidth={2}
+               label={''}
+               tooltip={'select a coordinate'}
+           />
+        </div>
     );
 
 }
@@ -62,7 +56,8 @@ WebGridUI.propTypes= {
 function onCoordinateChange(plotId, drawLayer, ev) {
     var csysName = get(ev, 'target.value');
     AppDataCntlr.dispatchAddPreference(COORDIANTE_PREFERENCE,csysName);
-    dispatchForceDrawLayerUpdate(drawLayer.displayGroupId, plotId);
+    dispatchModifyCustomField(drawLayer.displayGroupId,{COORDIANTE_PREFERENCE:csysName}, plotId);
+    //dispatchForceDrawLayerUpdate(drawLayer.displayGroupId, plotId);
 
 }
